refactor: :truck: change component structure
move component/common/layout/main-layout -> component/main-layout
