+++ src/asset/colors/activity-bar.ts
