+++ src/asset/icons/source-control.tsx
@@ -1,4 +1,4 @@
-import { WHITE } from "@src/component/common/atom/colors";
+import { WHITE } from "@src/asset/colors";
 import { IconProps } from ".";
 
 function SourceControlIcon({ style, fill = WHITE }: IconProps) {
