+++ src/component/main-layout/g-activity-bar/action-item.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { ActivityBar, TRANSPARENT } from "../../../atom/colors";
+import { ActivityBar, TRANSPARENT } from "../../../asset/colors";
 
 import { ActionItemType } from ".";
