+++ src/component/main-layout/g-footer.tsx
