+++ src/component/main-layout/g-sidebar.tsx
@@ -1,28 +1,23 @@
 import React from "react";
 import styled from "styled-components";
-import Link from "next/link";
 
-import { ACTION_ITEMS } from "./g-activity-bar";
-import { PostSlugType } from "@src/api/posts";
 import { useAppContext } from "@src/contexts/app";
 
-export type GSideBarProps = {
-  postSlugs: PostSlugType[];
-};
+import { ACTION_ITEMS } from "./g-activity-bar";
 
-function GSideBar({ postSlugs }: GSideBarProps) {
+function GSideBar() {
   const {
     state: { selectedActionItem },
   } = useAppContext();
 
+  const { Content } = ACTION_ITEMS[selectedActionItem];
+
   return (
     <Wrapper>
       <Title>{ACTION_ITEMS[selectedActionItem].label}</Title>
-      {postSlugs?.map(({ slug }) => (
-        <li key={slug}>
-          <Link href={`/posts/${slug}`}>{slug}</Link>
-        </li>
-      ))}
+      <ContentWrapper>
+        <Content />
+      </ContentWrapper>
     </Wrapper>
   );
 }
@@ -30,6 +25,9 @@ function GSideBar({ postSlugs }: GSideBarProps) {
 export default GSideBar;
 
 const Wrapper = styled.section`
+  display: flex;
+  flex-direction: column;
+
   width: 20rem;
   background-color: rgb(37, 37, 38);
   color: white;
@@ -40,4 +38,8 @@ const Title = styled.h2`
   padding: 0 0.8rem;
   font-weight: 400;
   opacity: 0.6;
+`;
+
+const ContentWrapper = styled.div`
+  flex: 1;
 `;
