+++ src/contexts/app/IAppContext.ts
@@ -1,5 +1,8 @@
+import { PostSlugType } from "@src/api/posts";
+
 export interface IAppContext {
   state: {
+    postSlugs: PostSlugType[];
     selectedActionItem: number;
   };
   action: {
