+++ src/contexts/app/index.tsx
@@ -1,5 +1,6 @@
 import { useState, createContext, useContext } from "react";
 
+import { PostSlugType } from "@src/api/posts";
 import { IAppContext } from "./IAppContext";
 
 const AppContext = createContext<IAppContext>(undefined!);
@@ -8,13 +9,18 @@ const useAppContext = () => useContext(AppContext);
 
 type AppContextProviderProps = {
   children: React.ReactNode;
+  postSlugs: PostSlugType[];
 };
 
-const AppContextProvider = ({ children }: AppContextProviderProps) => {
+const AppContextProvider = ({
+  children,
+  postSlugs,
+}: AppContextProviderProps) => {
   const [selectedActionItem, setSelectedActionItem] = useState<number>(0);
 
   const appStore: IAppContext = {
     state: {
+      postSlugs,
       selectedActionItem,
     },
     action: {
